don't use auto in lambda functions
this is unsupported in gcc-4.8. while this is uglier to read it fixes
compilation for older gcc versions that support c++1y, but not c++14.

an example platform that is affected is amazon linux 1, which is used in
aws lambda functions. therefore, this fix allows pysupercluster to be
compiled and used in aws lambda functions.